added pre-conversion checks
I added two functions to first check whether a given URL was not empty, and another function to check whether it's not already a png file.

1. Check if url length >0 to be true
2. pass the file extension into method, and check: if method == "png" --> true

Still want to add checks for whether url is not empty but isn't valid.
And mayeb add like a small window or box that slides to the right of the url box.,

And must add actual conversion