Handling Reddit URL
I created a method to handle the latest URL versions of reddit. (it somehow changed?)

Just searching for a keyword within each link, and then alerting the user (will remove later, probably).

Also trying to extend the "isPNG" method check, by checking if the given url filetype is the same as the desired filetype. But running into issues with reddit links, sigh...