Small fix: Reorganized filename,URL processing
I just reorganized the processing of the URL into the new filename.
I placed it at the top-most part, still inside the block of the convert() method.

This was done to fix the logic error, when calling the method to check if the current file extensions is equal to the desired one.